refactor prior to subclassing
