+++ src/zenml/integrations/evidently/steps/__init__.py
@@ -12,27 +12,27 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-from zenml.integrations.evidently.steps.evidently_categorical_target_drift import (
+from zenml.integrations.evidently.steps.profiles.evidently_categorical_target_drift import (
     EvidentlyCategoricalTargetDriftDetectionConfig,
     EvidentlyCategoricalTargetDriftDetectionStep,
 )
-from zenml.integrations.evidently.steps.evidently_classification_model_performance import (
+from zenml.integrations.evidently.steps.profiles.evidently_classification_model_performance import (
     EvidentlyClassificationModelPerformanceConfig,
     EvidentlyClassificationModelPerformanceStep,
 )
-from zenml.integrations.evidently.steps.evidently_data_drift import (
+from zenml.integrations.evidently.steps.profiles.evidently_data_drift import (
     EvidentlyDriftDetectionConfig,
     EvidentlyDriftDetectionStep,
 )
-from zenml.integrations.evidently.steps.evidently_numerical_target_drift import (
+from zenml.integrations.evidently.steps.profiles.evidently_numerical_target_drift import (
     EvidentlyNumericalTargetDriftDetectionConfig,
     EvidentlyNumericalTargetDriftDetectionStep,
 )
-from zenml.integrations.evidently.steps.evidently_probabilistic_model_performance import (
+from zenml.integrations.evidently.steps.profiles.evidently_probabilistic_model_performance import (
     EvidentlyProbabilisticModelPerformanceConfig,
     EvidentlyProbabilisticModelPerformanceStep,
 )
-from zenml.integrations.evidently.steps.evidently_regression_model_performance import (
+from zenml.integrations.evidently.steps.profiles.evidently_regression_model_performance import (
     EvidentlyRegressionModelPerformanceConfig,
     EvidentlyRegressionModelPerformanceStep,
 )
